# Faster video inference script.