@DaDaDaDaDaYeah 2023-07-02
Tested. This really works! Thanks!!

Test results (upscale parameter 2):
From 5-7 frame/s (original codes) to average 30 frame/s. Same output quality.
From GPU usage 30% (mostly CPU) to GPU usage 95-100%.