@tthg119 2023-07-03
Hmm, I have no idea why my result stays the same somehow. My run was `ESRNet_4xplus, nproc=1, [480x270] --> [1920x1080]` on 4x A40 (48Gb VRAM each), 52 cpu cores with 920G RAM. Both the old and new inference script give ~3fr/s
![Screenshot from 2023-07-03 09-56-55](https://github.com/xinntao/Real-ESRGAN/assets/75705305/bab18a1d-7927-42d0-9bd5-b9f8e57daed0)

Yes I noticed that the batch approach boost up GPU utilization significantly (around 23Gb 100% on each GPU comparing to just 4Gb 60% ish). I didn't measure cpu in details but Htop shows that it's about the same.
Also I tried with different models and configs, with different batch size but again the difference is only ~0.5fr/s. Would love to have some thoughts if possible