@FNsi 2023-07-13
25% faster for me!