@FNsi 2023-07-13
May I ask a question? 
Do you know why without --fp32 the output will be white noise? (Amd Rocm)