@FNsi 2023-07-13
> I am running the animevideov3 model without FP32 and the outputs are correct.

Sorry I tried with or without fp32 and there's no difference, whole white output at all.

> Could you please provide more details about your setup?

Just simply using Python inference.py --fp32 or not with general x3v4 model (the tiny denoise one)