@eliphatfs 2023-07-13
This command is working fine on my machine:
```bash
python inference_realesrgan_video_fast.py --model_name=realesr-general-x4v3 -i "videos\2022-12-24 17-53-30.mp4" -s 2
```

Did I understand your input correctly?