@FNsi 2023-07-13
> Did I understand your input correctly?

I think you are right. I did it with -dn 0, I will try to use it again without it.