@eliphatfs 2023-07-21
> FYI, I observe `torch.compile + channel_last` provides 2x speedup (no tiling, no face enhancing, fp16) on NVIDIA A4000.
> 
> ```
> self.model = self.model.to(memory_format=torch.channels_last)
> self.model = torch.compile(self.model)
> ```
> 
> Might be worth a shot?
> 
> I'm not sure how easy `torch.compile` fits with other features (face enhance, tiling), nor the end-to-end speedup in these cases (tiling and face enhace add CPU overheads).

Thanks for your comments! Which model are you using? On my side, using channel_last seems to _reduce_ performance by half.
`torch.compile` is generally helpful for performance as it generates optimized code on kernel level.