@eliphatfs 2023-07-21
https://github.com/pytorch/pytorch/issues/92542
I guess RRDB-based networks and VGG-based networks have different preferences for channel formats.