@aliencaocao 2024-05-22
how to use this on images instead of videos?